+++ nam/train/gui/__init__.py
@@ -10,72 +10,6 @@ Usage:
 >>> run()
 """
 
-
-# Hack to recover graceful shutdowns in Windows.
-# This has to happen ASAP
-# See:
-# https://github.com/sdatkinson/neural-amp-modeler/issues/105
-# https://stackoverflow.com/a/44822794
-def _ensure_graceful_shutdowns():
-    import os
-
-    if os.name == "nt":  # OS is Windows
-        os.environ["FOR_DISABLE_CONSOLE_CTRL_HANDLER"] = "1"
-
-
-_ensure_graceful_shutdowns()
-
-
-def _apply_extensions():
-    def removesuffix(s: str, suffix: str) -> str:
-        # Remove once 3.8 is dropped
-        if len(suffix) == 0:
-            return s
-        return s[: -len(suffix)] if s.endswith(suffix) else s
-
-    import importlib
-    import os
-    import sys
-
-    # DRY: Make sure this matches the test!
-    home_path = os.environ["HOMEPATH"] if os.name == "nt" else os.environ["HOME"]
-    extensions_path = os.path.join(
-        home_path, ".neural-amp-modeler", "extensions"
-    )
-    if not os.path.exists(extensions_path):
-        return
-    if not os.path.isdir(extensions_path):
-        print(
-            f"WARNING: non-directory object found at expected extensions path {extensions_path}; skip"
-        )
-    print("Applying extensions...")
-    if extensions_path not in sys.path:
-        sys.path.append(extensions_path)
-        extensions_path_not_in_sys_path = True
-    else:
-        extensions_path_not_in_sys_path = False
-    for name in os.listdir(extensions_path):
-        if name in {"__pycache__", ".DS_Store"}:
-            continue
-        try:
-            importlib.import_module(removesuffix(name, ".py"))  # Runs it
-            print(f"  {name} [SUCCESS]")
-        except Exception as e:
-            print(f"  {name} [FAILED]")
-            print(e)
-    if extensions_path_not_in_sys_path:
-        for i, p in enumerate(sys.path):
-            if p == extensions_path:
-                sys.path = sys.path[:i] + sys.path[i + 1 :]
-                break
-        else:
-            raise RuntimeError("Failed to remove extensions path from sys.path?")
-    print("Done!")
-
-
-_apply_extensions()
-
-
 import re
 import tkinter as tk
 import sys
